fix(#730): break org-context/actor-context/api-key circular dependency
ActorType/ActorContext move to a new actor-types.ts. org-context.ts,
request-actor.ts, and api-key.ts now import the type from there instead of
from actor-context.ts — actor-context.ts's getSessionActorContext() calls
org-context.ts's getOrgContext(), so importing actor-context.ts from any of
those three closed a cycle back to org-context.ts. actor-context.ts re-exports
the types so its test and other consumers are unaffected.

POLICY.md R-3.5.
